introduced a new registration material controller for my new method and moved files arounf in tests
