+++ src/EPR.PRN.Backend.API.UnitTests/Controllers/Regulator/AccreditationMaterialControllerTests.cs
@@ -15,7 +15,7 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using EPR.PRN.Backend.API.Common.Enums;
 
-namespace EPR.PRN.Backend.API.UnitTests.Controllers;
+namespace EPR.PRN.Backend.API.UnitTests.Controllers.Regulator;
 
 [TestClass]
 public class AccreditationMaterialControllerTests
