+++ src/EPR.PRN.Backend.API.UnitTests/Controllers/Regulator/RegulatorApplicationTaskStatusControllerTests.cs
@@ -11,37 +11,37 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using System.Net;
 
-namespace EPR.PRN.Backend.API.UnitTests.Controllers;
+namespace EPR.PRN.Backend.API.UnitTests.Controllers.Regulator;
 
 [TestClass]
-public class RegulatorRegistrationTaskStatusControllerTests
+public class RegulatorApplicationTaskStatusControllerTests
 {
-    private RegulatorRegistrationTaskStatusController _systemUnderTest;
+    private RegulatorApplicationTaskStatusController _systemUnderTest;
 
     private Mock<IMediator> _mockMediator;
-    private Mock<ILogger<RegulatorRegistrationTaskStatusController>> _mockLogger;
-    private Mock<IValidator<UpdateRegulatorRegistrationTaskCommand>> _updateRegulatorRegistrationTaskCommandValidatorMock;
+    private Mock<ILogger<RegulatorApplicationTaskStatusController>> _mockLogger;
+    private Mock<IValidator<UpdateRegulatorApplicationTaskCommand>> _updateRegulatorApplicationTaskCommandValidatorMock;
     private static readonly IFixture _fixture = new Fixture();
 
     [TestInitialize]
     public void TestInitialize()
     {
         _mockMediator = new Mock<IMediator>();
-        _mockLogger = new Mock<ILogger<RegulatorRegistrationTaskStatusController>>();
+        _mockLogger = new Mock<ILogger<RegulatorApplicationTaskStatusController>>();
 
-        _updateRegulatorRegistrationTaskCommandValidatorMock = new();
+        _updateRegulatorApplicationTaskCommandValidatorMock = new();
 
-        _systemUnderTest = new RegulatorRegistrationTaskStatusController(_mockMediator.Object, _updateRegulatorRegistrationTaskCommandValidatorMock.Object, _mockLogger.Object);
+        _systemUnderTest = new RegulatorApplicationTaskStatusController(_mockMediator.Object, _updateRegulatorApplicationTaskCommandValidatorMock.Object, _mockLogger.Object);
     }
 
     [TestMethod]
     public async Task Patch_RegulatorApplicationTaskStatus_ReturnsOk_WhenValidUpdateTaskStatusRequestDto()
     {
         //Arrange
-        var expectedTaskStatus = _fixture.Create<UpdateRegulatorRegistrationTaskCommand>();
+        var expectedTaskStatus = _fixture.Create<UpdateRegulatorApplicationTaskCommand>();
 
         var validationResult = new ValidationResult();
-        _updateRegulatorRegistrationTaskCommandValidatorMock.Setup(x => x.Validate(It.IsAny<UpdateRegulatorRegistrationTaskCommand>())).Returns(validationResult);
+        _updateRegulatorApplicationTaskCommandValidatorMock.Setup(x => x.Validate(It.IsAny<UpdateRegulatorApplicationTaskCommand>())).Returns(validationResult);
 
         //Act
         var result = await _systemUnderTest.UpdateRegistrationTaskStatus(expectedTaskStatus);
@@ -52,18 +52,18 @@ public class RegulatorRegistrationTaskStatusControllerTests
     }
 
     [TestMethod]
-    public async Task Patch_RegulatorRegistrationTaskStatus_ThrowsValidationException_WhenValidationFails()
+    public async Task Patch_RegulatorApplicationTaskStatus_ThrowsValidationException_WhenValidationFails()
     {
         // Arrange
-        var validator = new InlineValidator<UpdateRegulatorRegistrationTaskCommand>();
+        var validator = new InlineValidator<UpdateRegulatorApplicationTaskCommand>();
         validator.RuleFor(x => x.Status).Must(_ => false).WithMessage("Validation failed");
 
-        _systemUnderTest = new RegulatorRegistrationTaskStatusController(_mockMediator.Object, validator, _mockLogger.Object);
+        _systemUnderTest = new RegulatorApplicationTaskStatusController(_mockMediator.Object, validator, _mockLogger.Object);
 
         var registrationMaterialId = Guid.Parse("a9421fc1-a912-42ee-85a5-3e06408759a9");
-        var requestDto = new UpdateRegulatorRegistrationTaskCommand
+        var requestDto = new UpdateRegulatorApplicationTaskCommand
         {
-            RegistrationId = registrationMaterialId,
+            RegistrationMaterialId = registrationMaterialId,
             TaskName = "Test Task",
             Status = (RegulatorTaskStatus)999,
             UserName = "UserName"
@@ -79,8 +79,8 @@ public class RegulatorRegistrationTaskStatusControllerTests
     public async Task Patch_RegulatorApplicationTaskStatus_ThrowsException_WhenMediatorThrowsException()
     {
         // Arrange
-        var expectedTaskStatus = _fixture.Create<UpdateRegulatorRegistrationTaskCommand>();
-        _mockMediator.Setup(m => m.Send(It.IsAny<UpdateRegulatorRegistrationTaskCommand>(), It.IsAny<CancellationToken>())).ThrowsAsync(new Exception("Test exception"));
+        var expectedTaskStatus = _fixture.Create<UpdateRegulatorApplicationTaskCommand>();
+        _mockMediator.Setup(m => m.Send(It.IsAny<UpdateRegulatorApplicationTaskCommand>(), It.IsAny<CancellationToken>())).ThrowsAsync(new Exception("Test exception"));
 
         // Act
         Func<Task> act = async () => await _systemUnderTest.UpdateRegistrationTaskStatus(expectedTaskStatus);
@@ -89,62 +89,59 @@ public class RegulatorRegistrationTaskStatusControllerTests
         await act.Should().ThrowAsync<Exception>().WithMessage("Test exception");
     }
     [TestMethod]
-    public async Task AddRegistrationTaskQueryNote_ReturnsNoContent_WhenValidInput()
+    public async Task AddApplicationTaskQueryNote_ReturnsNoContent_WhenValidInput()
     {
         // Arrange
-        var validator = new InlineValidator<UpdateRegulatorRegistrationTaskCommand>();
+        var validator = new InlineValidator<UpdateRegulatorApplicationTaskCommand>();
         var taskId = Guid.NewGuid();
-        var command = new AddRegistrationTaskQueryNoteCommand
+        var command = new AddApplicationTaskQueryNoteCommand
         {
             Note = "This is a valid note",
             CreatedBy = Guid.NewGuid()
         };
 
         // Act
-        _systemUnderTest = new RegulatorRegistrationTaskStatusController(_mockMediator.Object, validator, _mockLogger.Object);
-        var result = await _systemUnderTest.RegistrationTaskQueryNote(taskId, command);
+        _systemUnderTest = new RegulatorApplicationTaskStatusController(_mockMediator.Object, validator, _mockLogger.Object);
+        var result = await _systemUnderTest.AddApplicationTaskQueryNote(taskId, command);
 
         // Assert
         result.Should().BeOfType<NoContentResult>();
     }
 
     [TestMethod]
-    public async Task AddRegistrationTaskQueryNote_ReturnsBadRequest_WhenValidationFails()
+    public async Task AddApplicationTaskQueryNote_ReturnsBadRequest_WhenValidationFails()
     {
         // Arrange
         var taskId = Guid.Empty;
-        var command = new AddRegistrationTaskQueryNoteCommand()
+        var command = new AddApplicationTaskQueryNoteCommand
         {
             Note = "", // Invalid note
             CreatedBy = Guid.Empty // Invalid QueryBy
         };
 
-        // Assert
-        
-        Func<Task> act = () => _systemUnderTest.RegistrationTaskQueryNote(taskId, command);
-
+        // Act
+        Func<Task> act = () => _systemUnderTest.AddApplicationTaskQueryNote(taskId, command);
         await act.Should().ThrowAsync<ValidationException>();
     }
 
     [TestMethod]
-    public async Task AddRegistrationTaskQueryNote_ReturnsInternalServerError_WhenMediatorThrows()
+    public async Task AddApplicationTaskQueryNote_ReturnsInternalServerError_WhenMediatorThrows()
     {
         // Arrange
         var taskId = Guid.NewGuid();
-        var command = new AddRegistrationTaskQueryNoteCommand
+        var command = new AddApplicationTaskQueryNoteCommand
         {
             Note = "Some valid note",
             CreatedBy = Guid.NewGuid()
         };
 
-        _mockMediator.Setup(x => x.Send(It.IsAny<AddRegistrationTaskQueryNoteCommand>(), It.IsAny<CancellationToken>()))
+        _mockMediator.Setup(x => x.Send(It.IsAny<AddApplicationTaskQueryNoteCommand>(), It.IsAny<CancellationToken>()))
             .ThrowsAsync(new Exception("Unexpected failure"));
 
         // Act
-        Func<Task> act = async () => await _systemUnderTest.RegistrationTaskQueryNote(taskId, command);
+        Func<Task> act = async () => await _systemUnderTest.AddApplicationTaskQueryNote(taskId, command);
 
         // Assert
         await act.Should().ThrowAsync<Exception>().WithMessage("Unexpected failure");
     }
-
 }
