+++ src/EPR.PRN.Backend.sln.DotSettings.user
@@ -5,7 +5,12 @@
   &lt;/TestAncestor&gt;&#xD;
 &lt;/SessionState&gt;</s:String>
 	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=7d5f46c5_002Deda3_002D4084_002Daac2_002D386dcb77f152/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" IsActive="True" Name="UpdateAsync #2" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
-  &lt;TestAncestor&gt;&#xD;
-    &lt;TestId&gt;MSTest::E32724E4-48EB-47A2-8239-CCD9F466EF3D::net8.0::EPR.PRN.Backend.Data.UnitTests.Repositories.RegistrationRepositoryTests&lt;/TestId&gt;&#xD;
-  &lt;/TestAncestor&gt;&#xD;
+  &lt;Or&gt;&#xD;
+    &lt;TestAncestor&gt;&#xD;
+      &lt;TestId&gt;MSTest::E32724E4-48EB-47A2-8239-CCD9F466EF3D::net8.0::EPR.PRN.Backend.Data.UnitTests.Repositories.RegistrationRepositoryTests&lt;/TestId&gt;&#xD;
+      &lt;TestId&gt;MSTest::4EE4E46A-2437-468E-BEFE-12C8DE172B3E::net8.0::EPR.PRN.Backend.API.UnitTests.Controllers.RegistrationMaterialControllerTests&lt;/TestId&gt;&#xD;
+      &lt;TestId&gt;MSTest::E32724E4-48EB-47A2-8239-CCD9F466EF3D::net8.0::EPR.PRN.Backend.Data.UnitTests.Repositories.Regulator.RegistrationMaterialRepositoryTests&lt;/TestId&gt;&#xD;
+    &lt;/TestAncestor&gt;&#xD;
+    &lt;Project Location="C:\Repos\epr-prn-common-backend\src\EPR.PRN.Backend.Data.UnitTests" Presentation="&amp;lt;EPR.PRN.Backend.Data.UnitTests&amp;gt;" /&gt;&#xD;
+  &lt;/Or&gt;&#xD;
 &lt;/SessionState&gt;</s:String></wpf:ResourceDictionary>
